window update changes
I made some changes so that windows can update after transfers. Right
now, only the Checkings and Savings windows will update if you click on
their buttons in the main Window, and that one will also update once
either of those are clicked.

I'm trying to get the main Window to update whenever a transfer happens,
or when that window closes.